// backend/src/enums/database.enums.ts
export enum DeliveryType {
  LIGHT = 'LIGHT',
  HEAVY = 'HEAVY'
}

export enum UserRole {
  USER = 'USER',
  ADMIN = 'ADMIN'
}

export enum CustomerType {
  RETAIL = 'RETAIL',
  WHOLESALE = 'WHOLESALE'
}

export enum InvoiceStatus {
  UNPAID = 'UNPAID',
  PARTIALLY_PAID = 'PARTIALLY_PAID',
  PAID = 'PAID',
  OVERDUE = 'OVERDUE'
}

export enum VehicleType {
  TRUCK_1_5T = 'TRUCK_1_5T',
  TRUCK_5T = 'TRUCK_5T'
}